Post-merge fixes of merging purge-db4o
Move class FCPPluginClientTest from package freenet.node.fcp to
freenet.clients.fcp as the related classes were moved like that as well
in purge-db4o.

This is part of a series of commits to fix compiler issues as a result
of the merge 85be221b19882802806ca12614e96ad156f466d6.
The series is not complete yet, the last commit will tell its the last.