Post-merge fixes of merging purge-db4o into plugin-fcp-rewrite
Use imports from freenet.clients.fcp instead of freenet.node.fcp
as the related classes were moved in purge-db4o.
I've reviewed all plugin-fcp-rewrite classes to conform to this.

This is part of a series of commits to fix compiler issues as a result
of the merge 85be221b19882802806ca12614e96ad156f466d6.
The series is not complete yet, the last commit will tell its the last.