Finish post-merge fixes of merging purge-db4o into plugin-fcp-rewrite
Make the JavaDoc talk about the equivalent class
freenet.clients.fcp.PersistentRequestClient instead of
freenet.node.fcp.FCPClient:
The PersistentRequestClient is the official replacement for FCPClient,
see: https://bugs.freenetproject.org/view.php?id=6298

Notice: The part of the JavaDoc which stressed the difference between
FCPPluginClient and FCPClient has been removed since the names are not
similar anymore.

This the last commit of the series of commits to fix compiler issues as
a result
of the merge 85be221b19882802806ca12614e96ad156f466d6.