FCPPluginClient.toStringShort(): Move to better place
This function is only used in one place - for determining the thread
name of the message dispatcher. As the message dispatcher is a class
itself already, move it there. It can even be toString() instead of
toStringShort() there since it has no toString() yet.
Also, document why this uses a "short" version of
FCPPluginClient.toString() instead of the whole of it.