FCPPluginClient: Extract interface FCPPluginConnection
This is the begin of the branch FCPPluginClient-refactoring which aims
to achieve the following:
- Rename FCPPluginClient to FCPPluginConnection. The first step is
the hereby commited extraction of the interface FCPPluginConnection.
FCPPluginClient to  will be renamed to FCPPluginConnectionImpl in
following commits.
- Add a adapter implementation of FCPPluginConnection which has a send()
function which does not require a SendDirection parameter but always
sends in a specific direction. This will allow passing an instance of
such an object to  the server which always sends to the client, and vice
verse.
- Prevent the server from keeping a hard reference to the
FCPPluginConnection. The server is already required to not kep a hard
reference by the JavaDoc, for example at
FredPluginFCPMessageHandler.ServerSideMessageHandler. This will be done
by giving it an adapter implementation of FCPPluginConnection which
internally only stores the ID of the real connection. The adapter will
always use PluginRespirator.getFCPluginClientByID() (which is subject to
be renamed to getFCPPluginConnectionByID() or similar) whenever
access to the real FCPPluginConnectionImpl is needed.