FCPPluginConnection: Add getServerPluginName()
Already implemented in the interface's implementation FCPPluginClient. I
had previously not added it to the interface since I wanted to keep it
package-private and interfaces can only have public functions.
But I would prefer for the package's internal code to typify variables
by the interface and not by the implementation, so the interface needs
to have that function.