FCPPluginClientTracker: Adapt to "Client" -> "Connection" rename
Class contents are finished.
Class is still to be renamed to FCPPluginConnectionTracker.

Further included improvements:
- JavaDoc of closedConnectionsQueue now says "nulled weak references"
instead of "removed items" because thats what it does actually monitor
- JavaDoc of ConnectionWeakReference now says that it removes "values"
from the TreeMap instead of only talking about the type of the values.
This is to stress that we need the key of the values, which is said in
the following sentence
- JavaDoc of getConnection(): Fix "The ID of [...] getID()" -> "The
value of [...]"
- JavaDoc of RealRun(): Fix "Pools" -> "Polls"