FCPPluginClientTracker: Rename to FCPPluginConnectionTracker
Also improves JavaDoc of registerConnection(): The @link to
getConnection doesn't have to include the class name since it is in the
same class.

Notice: The modified areas in class FCPServer do not rename the own
internal member variables of it to match the "client" -> "connection"
rename. This is intended. Class FCPServer will be dealt with in a
dedicated commit series.