FCPServer: Adapt to "Client" -> "Connection" rename
Notice: The functions where the return type lass FCPPluginClient is not
replaced with its parent interface FCPPluginConnection are intentional.
The callers need features of the actua FCPPluginClient implementation
which are not visible in the interface. This will be less ugly once
FCPPluginClient is renamed to FCPPluginConnnectionImpl in a following
commit.

Additional included improvements:
- Make all touched functions final
- Fix bogus grammar at start of
createFCPPluginClientForNetworkedFCP() JavaDoc
- Fix createFCPPluginClient*() JavaDoc to not say "and {@link
it anymore" because the function still works, it just wont return the
particular client which has been garbage collected.
- Rename getPluginClient() to get...ByID() as we also use that
level of verbosity in the caller PluginRespirator