FCPPluginClient: Rename to FCPPluginConnectionImpl
- Notice: The touched files show that there are some instances where we
keep using the type FCPPluginConnectionImpl instead of the interface
FCPPluginConnection. I will deal with those in following commits. Also,
some cases might sense to keep as is, because they need internal
features of the implementation which don't make sense to expose in the
interface.
- FCPConnectionHandler.getPluginClient() JavaDoc is changed to not
linkify FCPConnectionHandler anymore because IMHO linkifying a class in
its own JavaDoc confuses the reader into believing he is reading the doc
of a different class.
. FCPPluginClientTest JavaDoc: Add missing "by"