FCPPluginConnection: Remove getServerPluginName()
The function was not being used in anything except JavaDoc. I doubt that
server or client authors will need the function: If you're a client, you
know who you have been connecting to because you cannot create a
connection WITHOUT the name. If you're a server, you know who you are.
I had merely added it because FCPPluginConnectionImpl contains a member
variable to store it for internal use.

I'm removing it because having the function would complexify
implementation of  FCPPluginConnectionImpl.SendToClientAdapter: It would
have to cache the server name to implement getServerPluginName() - it
could not use getConnection() inside since that could throw
IOException, which getServerPluginName() shouldn't throw.