FCPPluginConnectionImpl.SendToClientAdapter: Optimize
Store a WeakReference instead of always querying the connection from the
FCPPluginConnectionTracker. This will save us a O(N*LOG(N)) query from
the FCPPluginConnectionTracker TreeMap for each call of getConnection(),
and thus for each send() / sendSynchronous().

This means that the class' visibility can be trimmed to private since it
shall only be newly constructed from the FCPPluginConnectionImpl now,
otherwise the constructor is not guaranteed to work.
This also means that object identity of a FCPPluginConnection is
guaranteed: We will always use the same SendToClientAdapter stored in
the FCPluginConnectionImpl, because it only allows the outside to query
that one, not create fresh ones.