FCPServer.createFCPPluginConnectionForIntraNodeFCP(): Trim return type
Reducing this to the parent interface FCPPluginConnection will allow
setting the visibility of FCPPluginConnectionImpl to package private
since what this commit changes is IIRC the only place which uses
FCPPluginConnectionImpl in a non-package-private way.

- This was made possible by implementing the usage of
getDefaultSendDirectionAdapter() in this function. I had speculated that
this function would be called in a caller of the FCPServer
function different package, which is why this used to return the
*Impl. But now that we call it here already, the FCPServer doesn't need
to return the Impl.