PluginRespirator: Organize imports.
This partially completes the branch
FCPPluginConnection-default-SendDirection
to the point of making everything  which hands out
FCPPluginConnection to client/server application code use a
FCPPluginConnectionImpl.getDefaultSendDirectionAdapter().

It should compile and work.

What remains to be done is adapting the JavaDoc of all plugin FCP stuff
to:
- Mention the default send directions
- NOT demand the server anymore to not keep a strong reference to the
FCPPluginConnection: The send direction adapters implicitely also do the
job of only keeping a weak reference, so the server can now keep a
strong one.

This TODO applies to the JavaDoc of at least the following:
- FCPPluginConnection,
- FCPPluginConnectionImpl
- FredPluginFCPMessageHandler
- PluginRespirator.connectToOtherPlugin()
- PluginRespirator.getPluginConnectionByID()