ServerSideFCPMessageHandler: Allow keeping strong refs to connections
The FCPPluginConnections which we pass to the message handler have
recently been changed to be a
FCPPluginConnectionImpl.SendToClientAdapter, which
internally only keeps a WeakReference to the FCPPluginConnectionImpl.
Thus, the server may keep a strong reference to that wrapper
without preventing GC of the actual FCPPluginConnectionImpl.

Additional included JavaDoc improvements:
- The "Notice that there is no..." sentence of the last item in the
enumeration is moved to the front to structure the sentences to be
ordered as 1. what is wrong 2. how to deal with it
- Don't push people very much to implement "Ping": Say "you may" instead
of "you are encouraged to". It is also possible to have implicit pinging
for stuff which sends messages very frequently anyway.