ClientRequester: Keep track of latest success/failure Date
This marks the beginning of branch issue-6384.
Its goal shall be to fix part of what is demanded at
https://bugs.freenetproject.org/view.php?id=6384

Specifically, the branch shall implement two UI columns for
downloads and uploads: "Last progress" and "Last failure".
They shall show the Dates which are added in this commit.
What the branch will NOT implement is keeping track of failure reasons:
As far as I understand, class ClientRequester is the "root" class which
stores downloads/upload block progress, so it should be the place which
stores  such information. But it doesn't actually receive reasons for
success/failure in its callbacks. Thus, it would be more work to
implement than I can currently afford to spend on this.

As for this specific commit, the following is notable:
There already *is* storage for "last activity" in ClientRequest (and
another class whose name I don't remember right now), but IMHO that
class is more low level and thus storage should be in the class to which
this commit adds it. The branch will deal with propagating the
information down to all classes which need it so it can be displayed in
the UI.

Notice: We do increment serialVersionUID.

This is because in a recent IRC discussion (Freenode IRC
channel #freenet) toad said that changes to classes which implement
Serizalizable should be accompanied by incrementing their
serialVersionUID.

With regards to backwards compatibility for people who are already using
purge-db4o testing releases, he said that the download queue should
handle this safely in a way which causn't loss of downloads: It has a
backup of the base data to restart them, and the backup is stored in a
non-Java-serialization fileformat. It even includes the progress of the
downloads.
Incrementing the serialVersionUID should cause the node to discard the
copies of the objects stored using serialization and use the
non-serialization backup instead.
(For uploads, the backup mechanism  was not implemented by
toad yet. This is apparent if you look at the function
ClientRequest.restartFrom() which powers
ClientLayerPersister.readRequestFromRecoveryData().
The code which loads uploads will however at least detect that the
serialVersionUID has changed and discard the affected uploads.)