*RequestStatus: Propgate Date latestSuccess & latestFailure
Commit d6c9ee6841ba65e58008891e6ef87e1461869f72 added Dates
latestSuccess & latestFailure to class ClientRequester.

Commit ff7f341c869ee2434fc1e2735f23c94b6feb1160 added the same to
SplitfileProgressEvent and adapted ClientRequester implementations to
propagate the dates to generated SplitfileProgressEvent objects.

As far as I understand class RequestStatus which this commit deals with,
it serves the following purpose related to ClientRequester /
SplitfileProgressEvent: Class RequestStatusCache provides a cached copy
of the latest state of a ClientRequester as it was propagated by
SplitfileProgressEvent objects. This is cached as a RequestStatus object
which is then used by the UI. This is done instead of having the UI
query the state directly from ClientRequester for concurrency/peformance
issues, i.e. not having to wait for locks.
Thus, RequestStatus is changed by this commit to also be able to store
latestSuccess & latestFailure and expose them via getters, so the UI
can use them.

The callers of the RequestStatus constructors in child classes are
updated to pass through the dates.
Users of the constructors in foreign classes are also adapt to the new
signature. Those take the Dates from the SimpleProgressMessage which
was recently changed to take them from the SplitfileProgressEvent
and expose them via getters.