FCP SimpleProgressMessage: Propagate Date latestSuccess & latestFailure
This adds the date the request last succeeded or failed in fetching a
block to the FCP message, as a Unix timestamp.

It also adds getters for internal fred use, since the following commit
will need those to propagate the information from the contained
SplitfileProgressEvent to constructors of RequestStatus objects.