ClientRequest: Remove obsolete, now unused lastActivity tracking
The previous commits moved the tracking to ClientRequester, and this
code is unused now.

Notice: We do increment serialVersionUID.

This is because in a recent IRC discussion (Freenode IRC
channel #freenet) toad said that changes to classes which implement
Serizalizable should be accompanied by incrementing their
serialVersionUID.

With regards to backwards compatibility for people who are already using
purge-db4o testing releases, he said that the download queue should
handle this safely in a way which causn't loss of downloads: It has a
backup of the base data to restart them, and the backup is stored in a
non-Java-serialization fileformat. It even includes the progress of the
downloads.
Incrementing the serialVersionUID should cause the node to discard the
copies of the objects stored using serialization and use the
non-serialization backup instead.
(For uploads, the backup mechanism  was not implemented by
toad yet. This is apparent if you look at the function
ClientRequest.restartFrom() which powers
ClientLayerPersister.readRequestFromRecoveryData().
The code which loads uploads will however at least detect that the
serialVersionUID has changed and discard the affected uploads.)