QueueToadlet: Use new RequestStatus.getLastFailure() / getLastActivity()
The column "Last failure" is newly added. It is only visible in advanced
mode.
The column "Last activity" is renamed to "Last progress" in the L10n.

The internal root source of both these Dates is
ClientRequester.latestSuccess / .latestFailure; which
are being passed through to the RequestStatus by the code added in the
preceeding commits. Thus for an understanding how the Dates are
computed, check class ClientRequester.

German L10n is also included since I am German.

Notice:
- Fields.compare() is not used anymore since getLastActivity() now
returns a Date instead of long.