ClientRequest(er): Revert increment of serialVersionUID
Commits d6c9ee6841ba65e58008891e6ef87e1461869f72 and
1ae98a0b4bf0be432db9b517f50cae3e80b0728f incremented the
serialVersionUID because member variables of these classes had been
added and removed.

Since changing the serialVersionUID would cause the node to discard the
serialized data of downloads/uploads, it would have to restore the
requests from the non-serialization backup. This backup only includes
downloads, not uploads, so the uploads would be lost.
Because purge-db4o has been deployed for months as
testing-build1468-pre1, the loss of uploads could affect many users.

Thus, this commit fixes the loss of uploads for 1468-pre testers by
keeping the  serialVersionUID at the same value as it was at
1468-pre before this branch started.
This causes the node to load the existing old copies of the serialized
ClientRequest(er) objects.

The added fields ClientRequester.latestSuccess/latestFailure are
initialized to null by Java's deserializer. This is fine: The code which
uses those fields already expects them to possibly be null. The UI will
display "unknown"/"never" then.
The value of the removed field ClientRequest.lastActivity is just
discarded by the deserizalitation code of Java, which is also fine
since we don't need it anymore.