Document that the new plugin FCP API is unstable and might change
The places where the comment is added are chosen as follows:
- PluginRespirator.connectToOtherPlugin() since that is the one and only
entry point for client plugins to obtain a connection to a server plugin
- FredPluginFCPMessageHandler since that is the one and only entry point
for server plugins to receive client connections.

In other words: Since those two places cover all code paths which allow
clients/servers to innitiate communication, I think it is enough to add
the notice there.