Update WebOfTrust to build0015
build0014 includes a huge amount of work: fixes bugs, updates for
purge-db4o, and adds an event-based API. build0015 is a small build and
tests fix on top of it - it makes no functional changes. The minimum
version is set to it because earlier versions will be broken with
purge-db4o.