bug4490: Recommended version support in official plugins
Ensure that all plugins have both a min and recommended version if either is set. Ensure that we download the recommended version all the time.

I've chosen against asking the user to confirm; reopen a new bug if you think it's something we should have. This hasn't been tested as thoroughly as it should have.