Fix updating for official plugins whose JAR URI is USK
My test runs show the fred code which fetches the USK does not try
hard enough to get the latest edition. Thus, this commit alows
plugin specifications to use alwaysFetchLatestVersion() to use a
negative USK edition. The fetcher will try harder with negative
ones.

Once the plugin is loaded, the JAR is kept on disk across restarts.
The previous code would then realize that it has a cached JAR, and not
ever try fetching the USK again. This is fixed by
alwaysFetchLatestVersion() as well.