OfficialPlugins: Add testing versions of WOT published by xor
Loading of both WebOfTrust and WebOfTrustTesting is prevented:
- If WebOfTrust is already loaded, and the user tries to load
WebOfTrustTesting, line 1323 of PluginManager.java will abort since they
both have the same main class. It won't display an error message, which
is why the added l10n tells the user to unload WebOfTrust before trying
to load *Testing. Notice that it is good that there is no error message
and it looks like loading succeded: This prevents users from loading the
testing versions if they did not read the "ATTENTION" part of the l10n.
- If WebOfTrustTesting is loaded already, the plugins page of FProxy
interestingly doesn't even list WebOfTrust in the UI for loading
plugins. See PproxyToadlet.java line 367.

The USK is the same as of my flog in the default bookmarks, and
as of my WOT seed identity.