l10n: reduce configuration description aggression
Warnings about not to change a setting "unless you know what you are
doing" make the software seem hostile. Options are already in advanced
mode if they shouldn't typically be changed, and if someone who doesn't
know what they're doing is changing settings they are liable to do so
even if the description advises against it. If a setting really should
never be changed it should be read-only in the interface.