Eclipe classpath: Fix sourcepath of contrib to match repository name
The previous sourcepath was uppercase, while our contrib repository uses
a lowercase name:
    https://github.com/freenet/contrib
If people clone it, they will get a "contrib" directory, not
"Contrib" as the previous sourcepath used.

This commit changes the sourcepath to be lowercase just like the
repository name.

This fixes Eclipse to find the db4o source code automatically, which is
useful for Web of Trust and Freetalk developers: Eclipse will be able to
display db4o JavaDoc now.