Fix test: Wait until it is *actually* sleeping.
The code was doing "wait until it has slept". This is not useful if it sleeps
and then wakes up again; we can either set a flag when we are waiting for it
to sleep, or we can wait until it is actually idle as measured by sleeping for
the maximum wait time.