Re-check the stores list periodically
This could actually happen if we have delayed initialisation. E.g. we
start accessing the ordinary datastore, and start a push job. Then the
password is entered and we start accessing the client cache. But the
snapshot the push thread took doesn't include it. So size never becomes 0,
the blocks are never written, and it loops forever.