Simplify and prevent matching bogus 1b2a3b4
A b expression is just a sequence of numbers separated by 'b''s. It may
not include anything else, as is obvious from the verifier method itself.
So the old code could incorrectly match 1b2a3b4 as a single b (&&)
expression, and then fail in the verifier. Note also that if we use
more complex patterns, && binds more strongly than ||, so it would be
wrong to allow this even if we could support it.