Limit scanning of all peers for unmatched packets
If there are too many unmatched packets, then ignore packets which we
can't match quickly for a while. This only happens when our peers change
IP address and we are port forwarded, or when it really is garbage e.g.
old (incompatible) nodes trying to connect to a seednode. Hence we prevent
a CPU-based denial of service attack, but at the cost of not being able
to connect immediately to peers that change their IP address. We will
still be able to connect eventually e.g. by looking up the ARK. If there
is no attack, and your node's IP address has never been used by a
seednode, this should not change anything.