Explicitly skip entire ID3v1 and ID3v2 tags
Previously we would sometimes find valid frame syncs in ID3 tag data.
This fix ensures that when we encounter such a tag, we skip it entirely
before trying to find the next frame.

While ID3v1 TAG fields may only legally occur at the end of a file, and
ID3v2 tags may only occur at the start of a file, concatenation of
several audio files may result is stray tags in the inner parts of the
file. There's no reason to loose sync on such fields, with all the
nasty effects, given we can easily skip and ignore them.