Use tracker to check/wait for request completion
This should give reproducible results because each request will finish
before starting the next. This is true anyway for SSK requests, but CHKs
have an opennet-related (but not opennet-only) stage at the end. It
would be significantly more work/overhead to make realGetCHK return only
after that, e.g. by making UIDTag's themselves waitable (we only finish
unlocking when the FNPOpennetCompletedAck is acknowledged!).