Send the new format UOMAnnouncement message for legacy
Since 1421, we have a new message, UOMAnnouncement. Nodes updating from
versions between 1421 and 1470 will recognise this. It does not include
freenet-ext.jar, which was moved into the dependency mechanism. Apart
from allowing us to get rid of the old message, this means we don't
need to fetch freenet-ext.jar from the USK (which may very well fail,
because it hasn't been used since 1421!).