BaseL10n: Add support for Portgual's Portuguese translations
Needed for the currently in development WoT build0019.

I have not yet filed a bug for the FIXME which this commit adds:
We should first discuss whether to just ditch the whole of this
handmade language code list and replace it with one downloaded from
Transifex / ISO / IETF. That would ensure that violations of standard
lists don't happen due to developers making mistakes when adding codes
by hand- which  already did happen in the  existing handmade list with
pt-br.