BaseL10n: Add support for Portgual's Portuguese translations
Needed for the currently in development WoT build0019.

Notice: This does not match what the Transifex command line tool uses.
Thus, please update your settings at ".tx/config".
(It cannot do so because Transifex does not seem to adhere to any
standard with its "pt_PT" code. We want to adhere to the IETF tags as
specified in the JavaDoc.)