Replace traversal-only LinkedList and HashMap by ArrayList
Except for at construction of the MessageFilter, we never lookup field
names, but only traverse the list sequentially. This makes it
unnecessary to keep a HashMap with its associated complexity and memory
costs.

This implementation keeps track of the field names and values
separately in ArrayLists kept consistent, so we can inexpensively
iterate over both while slightly reducing memory costs.

The changes include some minor deduplication of matching code.