Reduce frequency of MessageFilter cleanup job
When run every 100ms, the timed-out removal job only removes a handful
of jobs on an average node, meanwhile blocking access to the filters
for (measured) 0.3% of the time. For busy high-bandwidth nodes, this
count can expected to be higher.

This change increases the minimum and maximum wait times for the
removal job tenfold. Note that this removal job is not the only place
where timed-out filters are removed: checkFilters will also eventually
remove all timed-out filters, so we can safely decrease the cleanup job
frequency.