Fix peer node network traffic statistics
PeerNode previously failed to increment in/out traffic statistics
since startup, and did not initialize the total counters to include
the counts from prior Freenet sessions. The counters intended to
count the in/out traffic statistics since startup were initialized
to the total count and never incremented. This means that the two
sets of counters were reversed from their named purpose. This patch
fixes this and changes the total counters into totals at startup time.
The all time total traffic can be calculated as the sum of the traffic
since startup and the traffic before startup. No external code is aware
of this internal bookkeeping change, but it eliminates the need to
actively maintain both counters during runtime.