Remove the high-level TokenBucket's from NodeStats.
These are an obsolete part of load management. Output bandwidth liability
is dominant in practice. These are just unnecessary complexity IMHO. They
do not improve performance and do not guarantee any sort of sanity for
new nodes, especially as they rely on the measured average number of
bytes sent, which could vary for a new node.