fcp: fix ReturnType.DIRECT NPE
For a BinaryBlobWriter (in its current form) to be constructed the
Bucket it is passed must not be null.

Via Sadao: In the ReturnType=direct mode the request was stalled at the
very beginning due to NPE:

java.lang.NullPointerException
    at freenet.client.async.BinaryBlobWriter.getOutputStream(BinaryBlobWriter.java:74)
    at freenet.client.async.BinaryBlobWriter.addKey(BinaryBlobWriter.java:96)
    at freenet.client.async.ClientGetter.addKeyToBinaryBlob(ClientGetter.java:764)
    at freenet.client.async.SingleFileFetcher.onSuccess(SingleFileFetcher.java:193)

Introduced by 16d581ee416f01ce0136cf829bae4aad2f461cf6.