Route randomly for the first few hops.
This should improve performance, especially for bulk inserts. It should
also improve anonymity. It will increase request latency, and therefore
is disabled for realtime requests for now, but it should increase
throughput significantly, especially on darknet. Random routing for more
than one hop makes sense because our peers' capacity-to-random-keyspace
may be similar to ours given clustering. We random route at HTL 18 (2
hops) for bulk requests and at 18 and 17 (3 hops) for bulk inserts. Note
that Freenet 0.5 did this, but I didn't reinstate it because I didn't
understand it at the time.