BaseL10n: Begin changing the language code standard we use
This branch aims to add new language codes which are not contained in
ISO639 which the JavaDoc recommended previously.
To be able to add those codes in a standards-compliant way, we decided
to hereby change the standard we use to RFC5646.

For a lengthy discussion about this, see:
    https://github.com/freenet/fred/pull/500

Preexisting language codes which don't adhere to RFC5646 are only marked
as TODO, not renamed directly, because:
There are translations for them on Transifex already so it is likely
that plugins already use the existing codes in l10n file names.
Renaming the languages now would break those translations.
As we plan to move to a pre-defined RFC5646 list someday, we will likely
have to break many filename mappings then. So it's better to do this all
at once in the future than breaking some now and breaking some more
then.