Make KeyListenerTracker more resilient
This is a light version of c15b24d095205177cc959f825ced4f07e308de9a
with only the minimal required changes.

This adds a try/catch(Throwable) around all calls to KeyListener
methods that are inside a loop. Methods handling single KeyListeners
are not made more resilient, it makes sense to have the caller handle
errors in this case.