Cobertura: Fix Ant error if disabled
Travis CI would fail with:
 BUILD FAILED
 /home/travis/build/freenet/fred/build-clean.xml:407: Problem: failed to create task or type cobertura-report
 Cause: The name is undefined.
 Action: Check the spelling.
 Action: Check that any custom tasks/types have been declared.
 Action: Check that any <presetdef>/<macrodef> declarations have taken place.

This happened because the Cobertura Ant code is only loaded if the
"unit-cobertura-inject" target is run.
Fix by wrapping the affected code in a target of its own so Ant doesn't
try to figure out what "cobertura-report" is unless we're sure that
Cobertura has been loaded.