Force stripping ECDSA signature padding: BC 1.54 has retarded checks
Starting from BC 1.54 [0], ECDSA signatures are checked on whether they
occupy the entire byte array. If they do not, an exception is thrown.

This breaks the vast majority of JFK4 messages for users where BC
provides the ECDSA functionality. SunEC parses the padded signatures
just fine, as did BC 1.53.

As a workaround, parse the DER-encoded signature length and adjust the
signature length accordingly, effectively stripping the padding.

Thanks a lot, BouncyCastle developers, for introducing this far from
subtle, non-standard and completely unnecessary non-feature, and for
not mentioning it in your release notes.

[0] Introduced in d62b2f172e937dcbe8dd0a1ea614d24b2db88856, see
    https://github.com/bcgit/bc-java/