BookmarkItem: Remove unused constructor...
... so I don't need to update it in upcoming code which will add a new
parameter to all constructors.

From the Git history it looks like this was part of a previous on-disk
fileformat and forgotten to be removed.

I don't think plugins should be accesing this code directly:
BookmarkManager has functions for creating bookmarks, those should
be preferable to plugins.
So I'd say we can just modify the constructors as we please; and thus
remove this one without deprecating it first.