Ant builder: Force user to notice deprecation in favor of Gradle
Intended as a replacement for https://github.com/freenet/fred/pull/590

The Ant builder supports various things which the Gradle one cannot do
yet, for example:
- Building the JavaDoc
- Computing test coverage (https://github.com/freenet/fred/pull/553)
- Static code analysis using PMD (https://pmd.github.io/)
- Static code analysis using findbugs (http://findbugs.sourceforge.net/)
- Finding duplicate code with CPD
  (http://pmd.sourceforge.net/pmd-4.3.0/cpd.html)
- The Eclipse IDE (used by toad_ and xor)

I would thus really like to keep it.
To allow that, this commit will make the Ant builder fail with a
deprecation warning until the user disables the warning by command line
/ changing the override file.

As a compensation for keeping it in the codebase I will continue
maintaining it.