Revert part of commit 813c2c94c95e3cde72c7c2ba5a12bbb3ea2e3e9b
Don't make the ContentFilterTest expect the l10n to not work in unit
tests: It does only not work with the new experimental Gradle builder,
it does work with Ant. Thus it would fail with Ant, which this commit
fixes.
The next commit will fix it to work with Gradle.