gradle: fix typo in wrapper verification configuration
When the Gradle wrapper is downloaded it can be subject to SHA256
checksum verification provided that `distributionSha256Sum` is set in
`gradle-wrapper.properties`. A failure prints a message like this:

> Verification of Gradle distribution failed!
>
> Your Gradle distribution may have been tampered with.
> Confirm that the 'distributionSha256Sum' property in your gradle-wrapper.properties file is correct and you are downloading the wrapper from a trusted source.
>
>  Distribution Url: https://services.gradle.org/distributions/gradle-3.2-bin.zip
> Download Location: /home/steve/.gradle/wrapper/dists/gradle-3.2-bin/avantb235jk77byagvnd7aog7/gradle-3.2-bin.zip
> Expected checksum: '5321b36837226dc0377047a328f12010f42c7bf88ee4a3b1cee0c1104008293f'
>   Actual checksum: '5321b36837226dc0377047a328f12010f42c7bf88ee4a3b1cee0c11040082935'

However, if it is not set it does not perform this check. The existing
`gradle-wrapper.properties` did not, and set `distributionSha254Sum`
instead.