Travis CI: Fix for OpenJDK 7 by not using Gradle wrapper
Fixes these failures of the Gradle wrapper:
https://travis-ci.org/xor-freenet/fred-staging/jobs/286721010#L461-L511

This shows that freenet.crypt.KeyGenUtilsTest fails on OpenJDK 7:
https://travis-ci.org/xor-freenet/fred-staging/jobs/286768179#L536-L536

Notice: Instead of this PR we could also try updating the Gradle wrapper
binaries to see whether it fixes the failure. I'm not sure how I'm
supposed to obtain them so I did this PR as it does have some benefits
to not use the wrapper:
- Gradle is already installed in the Travis environment so it's faster
to not download it using the wrapper.
- We outsource the maintenance of figuring out which Gradle version to
use to be compatible with the rest of Travis.